+++ src/server/game/LuaEngine/LuaEngine.cpp
@@ -35,7 +35,6 @@ template<> const char* GetTName<Corpse>() { return "Corpse"; }
 template<> const char* GetTName<Weather>() { return "Weather"; }
 template<> const char* GetTName<AuctionHouseObject>() { return "AuctionHouse"; }
 
-extern void RegisterFunctions(lua_State* L);
 extern void AddElunaScripts();
 
 void StartEluna(bool restart)
